Vytváření receptu
... už funguje. Ještě potřebuje grafickej refactoring, ale jinak ok. Je
potřeba kompletní ROLLBACK A MIGRACE DB. Opravuju tam několik věcí, jako
třeba typ sloupce apod.
Opraven link na editaci receptu.

+++ src/rails/db/migrate/20121212143903_create_ingredience_recipe_connectors.rb
@@ -3,7 +3,7 @@ class CreateIngredienceRecipeConnectors < ActiveRecord::Migration
     create_table :ingredience_recipe_connectors do |t|
       t.references :ingredience
       t.references :recipe
-      t.integer :quantity,        :null => false, :default => 1
+      t.float :quantity,        :null => false, :default => 1
       t.integer :importance,      :null => false, :default => 1
 
       t.timestamps
